pdf-extractor-streamlit/frontend/requirements.txt
streamlit
streamlit-drawable-canvas
PyMuPDF
pillow
pandas
openpyxl
pytesseract
opencv-python
plotly
requests
numpy
